Refactor hustle market into modular lifecycle helpers
